backend/requirements.txt
Flask==3.1.0
Flask-Cors==4.0.0
gunicorn==22.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11
openpyxl==3.1.2
pandas==3.0.0
bcrypt==4.2.0
qrcode==8.2
jsonschema==4.19.0